+++ forwarder/forwarder_test.go
@@ -1,4 +1,4 @@
-package main
+package forwarder
 
 import (
 	"encoding/json"
