+++ forwarder/s3.go
@@ -1,4 +1,4 @@
-package main
+package forwarder
 
 import (
 	"fmt"
