+++ logcollector.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/Financial-Times/log-collector/forwarder"
+
+func main() {
+	forwarder.Forward()
+}
